Support ProxySQL 2.0.7 on CentOS 7
When `restart == true`, a systemd drop-in file is created to add the
`--reload` flag to the start command.  Previously, the module assumed
init.d only and the `start` parameter of the Service resource started
proxysql manually.  This resulted in the Pidfile not being written and
the systemd service reporting an incorrect status.

puppet/extlib is now required and camptocamp/systemd is a new soft-dependency.

Fixes #115